Done with adapters for RV and make little refactor

+++ app/src/main/java/com/eugenebaturov/gaymer/model/entities/Hero.kt
@@ -32,7 +32,7 @@ data class Hero(
     val baseMana: Int,
 
     @SerializedName(value = "base_armor")
-    val baseArmor: Int,
+    val baseArmor: Double,
 
     @SerializedName(value = "base_attack_min")
     val minDamage: Int,
